Underpass map file validation fix
In the latest versions of Rocket League, the original underpass map file exceeds the file size boundaries set in the 'isValidUnderpass' lambda function causing the application to throw an error every time a map is loaded. This change increases the boundaries so that this error isn't triggered and the maps can be loaded again.